Master

See merge request mes/mes-backend!38

+++ mes-modules/produce/src/main/java/com/richfit/mes/produce/controller/TrackAssignController.java
@@ -340,10 +340,12 @@ public class TrackAssignController extends BaseController {
                         if (CollectionUtils.isNotEmpty(ingredient.getLineList())) {
                             application = wmsServiceClient.anApplicationForm(ingredient).getData();
                         }
-                        if ("N".equals(application.getRetCode())) {
+                        //请勿重复上传！
+                        boolean upload = !application.getRetMsg().contains("请勿重复上传");
+                        if ("N".equals(application.getRetCode()) && upload) {
                             numberService.deleteApplicationNumberByItemId(trackItem.getId());
                             log.error("仓储数据:" + ingredient);
-                            throw new GlobalException("仓储服务:" + application.getRetMsg() + "加密数据:" + application.getEncryption(), ResultCode.FAILED);
+                            throw new GlobalException("仓储服务:" + application.getRetMsg(), ResultCode.FAILED);
                         }
                     }
                 }
